Refactor and cleanup according to review
