+++ examples/wallet/package.json
@@ -5,7 +5,6 @@
         "@concordium/web-sdk": "^7.0.3",
         "@scure/bip39": "^1.2.1",
         "buffer": "^6.0.3",
-        "jotai": "^2.5.1",
         "react": "18.2.0",
         "react-cookie": "^6.1.1",
         "react-dom": "^18.1.0",
