+++ examples/wallet/src/Identity.tsx
@@ -1,34 +1,8 @@
-import {
-    AttributeList,
-    CredentialDeploymentTransaction,
-    CredentialInput,
-    IdentityObjectV1,
-    getAccountAddress,
-    signCredentialTransaction,
-} from '@concordium/web-sdk';
-import React, { useEffect, useMemo, useState } from 'react';
-import { useLocation, useNavigate } from 'react-router-dom';
-import {
-    extractIdentityObjectUrl,
-    fetchIdentity,
-    getAccountSigningKey,
-    sendCredentialDeploymentTransaction,
-    getCryptographicParameters,
-    getIdentityProviders,
-    getDefaultTransactionExpiry,
-} from './util';
-import { mnemonicToSeedSync } from '@scure/bip39';
-import { Buffer } from 'buffer/';
-import { AccountWorkerInput } from './types';
-import {
-    credNumber,
-    identityIndex,
-    network,
-    seedPhraseKey,
-    selectedIdentityProviderKey,
-} from './constants';
-
-const worker = new Worker(new URL('./account-worker.ts', import.meta.url));
+import { AttributeList, IdentityObjectV1 } from '@concordium/web-sdk';
+import React, { useEffect, useState } from 'react';
+import { useLocation } from 'react-router-dom';
+import { extractIdentityObjectUrl, fetchIdentity } from './util';
+import { CreateAccount } from './CreateAccount';
 
 function DisplayIdentity({ attributes }: { attributes: AttributeList }) {
     return (
@@ -44,16 +18,8 @@ function DisplayIdentity({ attributes }: { attributes: AttributeList }) {
 
 export function Identity() {
     const location = useLocation();
-    const navigate = useNavigate();
     const [identity, setIdentity] = useState<IdentityObjectV1>();
     const [error, setError] = useState<string>();
-    const [createButtonDisabled, setCreateButtonDisabled] =
-        useState<boolean>(false);
-    const seedPhrase = useMemo(() => localStorage.getItem(seedPhraseKey), []);
-    const selectedIdentityProviderIdentity = useMemo(
-        () => localStorage.getItem(selectedIdentityProviderKey),
-        []
-    );
 
     useEffect(() => {
         const identityObjectUrl = extractIdentityObjectUrl(location.hash);
@@ -63,74 +29,6 @@ export function Identity() {
         // example.
         fetchIdentity(identityObjectUrl).then(setIdentity).catch(setError);
     }, [location.hash]);
-
-    async function createAndSendAccount() {
-        setCreateButtonDisabled(true);
-
-        if (
-            !identity ||
-            !seedPhrase ||
-            selectedIdentityProviderIdentity === null
-        ) {
-            setCreateButtonDisabled(false);
-            return;
-        }
-
-        const selectedIdentityProvider = (await getIdentityProviders()).find(
-            (idp) =>
-                idp.ipInfo.ipIdentity ===
-                Number.parseInt(selectedIdentityProviderIdentity)
-        );
-        if (!selectedIdentityProvider) {
-            setCreateButtonDisabled(false);
-            return;
-        }
-
-        const listener = (worker.onmessage = async (
-            e: MessageEvent<CredentialDeploymentTransaction>
-        ) => {
-            worker.removeEventListener('message', listener);
-            const credentialDeploymentTransaction = e.data;
-            const signingKey = getAccountSigningKey(
-                seedPhrase,
-                credentialDeploymentTransaction.unsignedCdi.ipIdentity
-            );
-            const signature = await signCredentialTransaction(
-                credentialDeploymentTransaction,
-                signingKey
-            );
-            await sendCredentialDeploymentTransaction(
-                credentialDeploymentTransaction,
-                signature
-            );
-            const accountAddress = getAccountAddress(
-                credentialDeploymentTransaction.unsignedCdi.credId
-            );
-            navigate(`/account/${accountAddress.address}`);
-        });
-
-        const global = await getCryptographicParameters();
-        const credentialInput: CredentialInput = {
-            net: network,
-            revealedAttributes: [],
-            seedAsHex: Buffer.from(mnemonicToSeedSync(seedPhrase)).toString(
-                'hex'
-            ),
-            idObject: identity,
-            identityIndex,
-            globalContext: global,
-            credNumber,
-            ipInfo: selectedIdentityProvider.ipInfo,
-            arsInfos: selectedIdentityProvider.arsInfos,
-        };
-        const expiry = getDefaultTransactionExpiry();
-        const workerInput: AccountWorkerInput = {
-            credentialInput,
-            expiry,
-        };
-
-        worker.postMessage(workerInput);
-    }
 
     if (error) {
         return (
@@ -147,12 +45,7 @@ export function Identity() {
             {identity && (
                 <>
                     <DisplayIdentity attributes={identity.attributeList} />
-                    <button
-                        disabled={createButtonDisabled}
-                        onClick={createAndSendAccount}
-                    >
-                        Create account
-                    </button>
+                    <CreateAccount identity={identity} />
                 </>
             )}
         </>
